getsym
